pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-triage-openenv"
version = "0.1.0"
description = "OpenEnv environment for customer support ticket triage and policy-safe resolution."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "openai>=1.50.0",
  "httpx>=0.27.0",
  "pyyaml>=6.0.2",
  "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["openenv_compat", "models", "tasks", "graders", "client"]
packages = ["server"]
